requirements.txt
python-jobspy>=1.1.0
pydantic>=2.7.0
PyYAML>=6.0.1
python-telegram-bot>=22.0
playwright>=1.44.0
playwright-stealth>=2.0.0
loguru>=0.7.2
tenacity>=8.3.0
python-dotenv>=1.0.1
pandas>=2.2.0
